+++ assets/src/scripts/monitoring-location/components/hydrograph/method-picker.test.js
@@ -1,11 +1,15 @@
 import {select} from 'd3-selection';
 
+import * as utils from 'ui/utils';
+
 import {configureStore} from 'ml/store';
 
+import * as dataIndicator from './data-indicator';
 import {drawMethodPicker} from './method-picker';
 import {TEST_PRIMARY_IV_DATA} from './mock-hydrograph-state';
 
 describe('monitoring-location/components/hydrograph/method-picker', () => {
+    utils.mediaQuery = jest.fn().mockReturnValue(true);
 
     describe('drawMethodPicker', () => {
         const TEST_STATE = {
@@ -17,8 +21,10 @@ describe('monitoring-location/components/hydrograph/method-picker', () => {
             }
         };
         let div;
+        let showDataIndicatorSpy;
         beforeEach(() => {
             div = select('body').append('div');
+            showDataIndicatorSpy = jest.spyOn(dataIndicator, 'showDataIndicators');
         });
 
         afterEach(() => {
@@ -47,7 +53,7 @@ describe('monitoring-location/components/hydrograph/method-picker', () => {
             expect(div.select('select').property('value')).toEqual('90649');
         });
 
-        it('selecting a different method updates the store', () => {
+        it('selecting a different method updates the store and updates the no data available indicator', () => {
             let store = configureStore(TEST_STATE);
             div.call(drawMethodPicker, store);
 
@@ -57,6 +63,8 @@ describe('monitoring-location/components/hydrograph/method-picker', () => {
             div.select('select').dispatch('change');
 
             expect(store.getState().hydrographState.selectedIVMethodID).toBe('90649');
+            expect(showDataIndicatorSpy.mock.calls).toHaveLength(1);
+            expect(showDataIndicatorSpy.mock.calls[0][0]).toBe(false);
         });
 
         it('Expects if the data has only one method then the picker will be hidden', () => {