+++ assets/src/scripts/monitoring-location/components/hydrograph/selectors/iv-data.test.js
@@ -43,11 +43,26 @@ describe('monitoring-location/components/hydrograph/selectors/iv-data', () => {
             expect(getIVDataPoints('compare')(TEST_STATE)).toBeNull();
         });
 
+        it('Returns an empty array if no data points exist for data kind', () => {
+            expect(getIVDataPoints('compare')({
+                ...TEST_STATE,
+                hydrographData: {
+                    ...TEST_STATE.hydrographData,
+                    compareIVData: {
+                        parameter: {
+                            parameterCode: '72019'
+                        },
+                        values: {}
+                    }
+                }
+            })).toHaveLength(0);
+        });
+
         it('Returns the iv data points with the correct properties', () => {
             const points = getIVDataPoints('primary')(TEST_STATE);
-            expect(points['90649']).toBeDefined();
-            expect(points['90649']).toHaveLength(9);
-            expect(points['90649'][0]).toEqual({
+            expect(points).toBeDefined();
+            expect(points).toHaveLength(9);
+            expect(points[0]).toEqual({
                 value: 24.2,
                 dateTime: 1582560900000,
                 isMasked: false,
@@ -56,7 +71,7 @@ describe('monitoring-location/components/hydrograph/selectors/iv-data', () => {
                 label: 'Approved',
                 class: 'approved'
             });
-            expect(points['90649'][2]).toEqual({
+            expect(points[2]).toEqual({
                 value: null,
                 dateTime: 1582562700000,
                 isMasked: true,
@@ -65,7 +80,7 @@ describe('monitoring-location/components/hydrograph/selectors/iv-data', () => {
                 label: 'Ice Affected',
                 class: 'ice-affected-mask'
             });
-            expect(points['90649'][4]).toEqual({
+            expect(points[4]).toEqual({
                 value: 25.2,
                 dateTime: 1582570800000,
                 isMasked: false,
@@ -74,7 +89,7 @@ describe('monitoring-location/components/hydrograph/selectors/iv-data', () => {
                 label: 'Estimated',
                 class: 'estimated'
             });
-            expect(points['90649'][7]).toEqual({
+            expect(points[7]).toEqual({
                 value: 26.5,
                 dateTime: 1600619400000,
                 isMasked: false,
@@ -129,10 +144,23 @@ describe('monitoring-location/components/hydrograph/selectors/iv-data', () => {
            expect(getIVDataSegments('compare')(TEST_STATE)).toBeNull();
        });
 
+       it('Expects an empty array if IV of data kind exists but there are no points', () => {
+           expect(getIVDataSegments('compare')({
+                ...TEST_STATE,
+                hydrographData: {
+                    ...TEST_STATE.hydrographData,
+                    compareIVData: {
+                        parameter: {
+                            parameterCode: '72019'
+                        },
+                        values: {}
+                    }
+                }
+            })).toHaveLength(0);
+       });
+
        it('Returns the expected data segments', () => {
-           const segmentsByMethodID = getIVDataSegments('primary')(TEST_STATE);
-           expect(segmentsByMethodID['90649']).toBeDefined();
-           const segments = segmentsByMethodID['90649'];
+           const segments = getIVDataSegments('primary')(TEST_STATE);
            expect(segments).toHaveLength(5);
            expect(segments[0]).toEqual({
                isMasked: false,
@@ -183,8 +211,23 @@ describe('monitoring-location/components/hydrograph/selectors/iv-data', () => {
     });
 
     describe('getIVUniqueDataKinds', () => {
-        it('returns an empty array if no IV data of data kind exists', () => {
-            expect(getIVUniqueDataKinds('compare')(TEST_STATE)).toHaveLength(0);
+        it('returns null if no IV data of data kind exists', () => {
+            expect(getIVUniqueDataKinds('compare')(TEST_STATE)).toBeNull();
+        });
+
+        it('returns an empty array if IV data of dataKind exists but has no points', () => {
+            expect(getIVUniqueDataKinds('compare')({
+                ...TEST_STATE,
+                hydrographData: {
+                    ...TEST_STATE.hydrographData,
+                    compareIVData: {
+                        parameter: {
+                            parameterCode: '72019'
+                        },
+                        values: {}
+                    }
+                }
+            })).toHaveLength(0);
         });
 
         it('Returns expected unique data kind for the IV data', () => {