+++ assets/src/scripts/monitoring-location/index.js
@@ -32,13 +32,15 @@ const load = function() {
         }
     });
     let nodes = document.getElementsByClassName('wdfn-component');
+    const hashOptions = Object.fromEntries(new window.URLSearchParams(getParamString()));
 
     for (let node of nodes) {
-        // If options is specified on the node, expect it to be a JSON string.
-        // Otherwise, use the dataset attributes as the component options.
-        const options = node.dataset.options ? JSON.parse(node.dataset.options) : node.dataset;
-        const hashOptions = Object.fromEntries(new window.URLSearchParams(getParamString()));
-        COMPONENTS[node.dataset.component](store, node, Object.assign({}, options, hashOptions));
+        if (!hashOptions.showOnlyGraph || node.dataset.component === 'hydrograph') {
+            // If options is specified on the node, expect it to be a JSON string.
+            // Otherwise, use the dataset attributes as the component options.
+            const options = node.dataset.options ? JSON.parse(node.dataset.options) : node.dataset;
+            COMPONENTS[node.dataset.component](store, node, Object.assign({}, options, hashOptions));
+        }
     }
 
     window.onresize = function() {