+++ CHANGELOG.md
@@ -5,7 +5,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/usgs/waterdataui/compare/waterdataui-0.43.0...master)
-
 ### Changed
 - The data loading indicator for the hydrograph now appears on top of the graph rather than with the date controls.
 - Cleaned up hydrograph rendering and made changes to make it possible to determine when rendering of the hydrograph is complete.
@@ -13,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 recent data is chosen. 
 - Floating feedback link is now centered on page (and networks page) and an additional link added to header.
 - Changed label order and label wording for the 'custom' option in time selection menu.
+
+### Added
+- If the graph shows no data for the selected time range, a message will appear on top of the graph. In addition, if a particular kind of selected data isn't available (for instance if compare to last year is selected and there is no data for that time range)
 
 ## [0.43.0](https://github.com/usgs/waterdataui/compare/waterdataui-0.43.0...waterdataui-0.42.0)
 ### Added